requirements.txt
Django==1.11.16
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1
argon2-cffi==18.3.0
auth==0.5.3
bcrypt==3.1.5
beautifulsoup4==4.6.0
bleach==2.1.4
blinker==1.4
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
defusedxml==0.5.0
django-allauth==0.38.0
django-appconf==1.0.2
django-autoslug-iplweb==1.9.4
django-bootstrap-form==3.4
django-ckeditor==5.6.1
django-classy-tags==0.8.0
django-currentuser==0.2.3
django-filter==2.0.0
django-fobi==0.13.7
django-formtools==2.1
django-friendship==1.8.0
django-haystack==2.8.1
django-js-asset==1.1.0
django-machina==0.7.1
django-modelcluster==4.3
django-mptt==0.9.1
django-nine==0.1.13
django-nonefield==0.3
django-nyt==1.1.2
django-paypal==0.5.0
django-sekizai==0.10.0
django-simple-history==2.6.0
django-taggit==0.23.0
django-treebeard==4.3
django-user-accounts==2.1.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
dnspython==1.16.0
draftjs-exporter==2.1.5
easy-thumbnails==2.5
eventlet==0.24.1
falcon==1.4.1
google-auth==1.6.2
greenlet==0.4.15
gunicorn==19.9.0
html5lib==1.0.1
idna==2.8
jsonfield==2.0.2
Markdown==2.6.11
markdown2==2.3.7
mongoengine==0.16.3
monotonic==1.5
oauthlib==2.1.0
passlib==1.7.1
Pillow==5.3.0
Pinax==0.9a2
pinax-eventlog==2.0.3
pinax-messages==2.0.2
pinax-templates==2.0.2
pinax-webanalytics==4.0.2
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyJWT==1.7.1
pymongo==3.7.2
python-mimeparse==1.6.0
python-social-auth==0.3.6
python3-openid==3.1.0
requests==2.21.0
requests-oauthlib==1.0.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
social-auth-core==2.0.0
sorl-thumbnail==12.5.0
Unidecode==1.0.23
urllib3==1.24.1
virtualenv==16.1.0
vishap==0.1.5
wagtail==2.3
webencodings==0.5.1
Werkzeug==0.14.1
wiki==0.4.2
Willow==1.1
